+++ templates/sidebar_home.inc.php
@@ -24,6 +24,7 @@
 $t_songs     = T_('Songs');
 $t_artists   = T_('Artists');
 $t_albums    = T_('Albums');
+$t_labels    = T_('Labels');
 $t_playlists = T_('Playlists');
 $t_tagcloud  = T_('Tag Cloud');
 $t_expander  = T_('Expand/Collapse');
@@ -46,7 +47,7 @@ $t_search    = T_('Search'); ?>
             <li id="sb_home_browse_music_album"><a href="<?php echo $web_path; ?>/mashup.php?action=album"><?php echo $t_albums; ?></a></li>
             <li id="sb_home_browse_music_artist"><a href="<?php echo $web_path; ?>/mashup.php?action=artist"><?php echo $t_artists; ?></a></li>
             <?php if (AmpConfig::get('label')) { ?>
-            <li id="sb_home_browse_music_label"><a href="<?php echo $web_path ?>/browse.php?action=label"><?php echo T_('Labels') ?></a></li>
+            <li id="sb_home_browse_music_label"><a href="<?php echo $web_path ?>/browse.php?action=label"><?php echo $t_labels ?></a></li>
             <?php
         } ?>
             <?php if (AmpConfig::get('channel')) { ?>
@@ -115,6 +116,10 @@ $t_search    = T_('Search'); ?>
           <li id="sb_home_search_song"><a href="<?php echo $web_path; ?>/search.php?type=song"><?php echo $t_songs; ?></a></li>
           <li id="sb_home_search_album"><a href="<?php echo $web_path; ?>/search.php?type=album"><?php echo $t_albums; ?></a></li>
           <li id="sb_home_search_artist"><a href="<?php echo $web_path; ?>/search.php?type=artist"><?php echo $t_artists; ?></a></li>
+          <?php if (AmpConfig::get('label')) { ?>
+          <li id="sb_home_search_label"><a href="<?php echo $web_path; ?>/search.php?type=label"><?php echo $t_labels; ?></a></li>
+                <?php
+            } ?>
           <li id="sb_home_search_playlist"><a href="<?php echo $web_path; ?>/search.php?type=playlist"><?php echo $t_playlists; ?></a></li>
           <?php if (AmpConfig::get('allow_video') && Video::get_item_count('Video')) { ?>
             <li id="sb_home_search_video"><a href="<?php echo $web_path ?>/search.php?type=video"><?php echo T_('Videos') ?></a></li>